refactor: 재사용을 위해 RefreshToken 도메인 및 레포지토리 공통 모듈로 이동
