+++ nmnb-application/src/main/kotlin/nmnb/application/global/auth/service/RefreshTokenService.kt
@@ -2,10 +2,10 @@ package nmnb.application.global.auth.service
 
 import nmnb.application.global.common.utils.DeviceIdUtils
 import nmnb.application.global.infrastructure.security.JwtProvider
+import nmnb.common.auth.RefreshToken
+import nmnb.common.auth.repository.RefreshTokenRepository
 import nmnb.common.response.exception.AuthException
 import nmnb.common.response.status.ErrorStatus
-import nmnb.domain.auth.RefreshToken
-import nmnb.domain.auth.repository.RefreshTokenRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
