+++ nmnb-common/src/main/kotlin/nmnb/common/auth/RefreshToken.kt
@@ -1,4 +1,4 @@
-package nmnb.domain.auth
+package nmnb.common.auth
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.redis.core.RedisHash
